:- BST.sol _ minor bug fix w/ checking for 0 amount before executing swap 	_exeTokBuyBurn now checks for 0 amount before proceeding to use uniswap 		swap or get amounts throws error on 0 amount 		now simply returns 0 and doesn’t attempt quote or swap 		this is needed because burning 0 is indeed a use case
:- _abi bug fix with func hash arg type
:- _web3 update default gas price ratio
:- _bst_keeper update with global DEBUG_LEVEL

deployed: # tBST37.4: 0x5FD6165a21dB9AfCFc61FbC64fa4c1f25854D9e4
NEXT: need python side testing (v37.4)
	- end-to-end testing: tg trinity <-> BST <-> web dapp (w/ multiple LPs on multiple dex routers)

+++ src/_abi.py
@@ -46,7 +46,7 @@ BST_TRADEIN_FUNC_SIGN = "tradeInBST(uint64)"
 BST_TRADEIN_FUNC_HASH = "d8785767"
 BST_FUNC_MAP_WRITE = {
     # write functions
-    "KEEPER_maintenance(uint256,address)": ["4dd534c0", ["uint64","address"], []], # gas used: 62,434
+    "KEEPER_maintenance(uint256,address)": ["4dd534c0", ["uint256","address"], []], # gas used: 62,434
     "KEEPER_setRatios(uint32,uint32)": ["3dcff192", ["uint32","uint32"], []], 
 
     "KEEPER_setKeeper(address)": ["11851737", ["address"], []],
